Add email and signing key variables back in
On my work laptop, I set this repository's local config on the machine
to specify the email and signing key. Now that I have a separate config
to work with enterprise repositories that is included to override the
email and signing key variables, I can add them back in my main git
configuration without worrying if they will mess my commits for work.