Set telescope border to true
Settings the border to false removes the border and the floating window
titles entirely. To get the borderless look, we need the border chars to
be empty or the foreground and background colors to be the same.

See the gallery in the telescope.nvim wiki. Or copy pasta the config
from Nvchad since that's the look that I'm going for.