WIP: plugin spec using Paq
Split up list of plugins into "start", "opt", and "colorscheme". Create
a helper function `opt_by_default`, which takes a list of plugins and
returns them specified as optional. Set `paq` as the metavalue for
`__index` in the module's metatable, so we call do stuff like
`require("plugins"):sync()`.

Yes, it is still very much a work in progress.