Use `keymaps` as module
Move from `plugin` to `lua` dir so we can assign keymaps as part of user
config initialization instead of plugin initialization. If we used
"--noplugin" when starting nvim, our keymaps would not get sourced.

Now, we can require the `keymaps` module from the init.lua to set
keymaps, regardless if we are going to load any plugins or not.