Export additional functions from the `map` module
Add a `traverse` helper function that uses coroutines!

Export functions for validating the table argument the `map` function.

Thinking about not validating the entire table when calling `map`...

Since having a huge-ass table containing a bunch of mappings will
probably be prone to incorrect formatting, maybe we can make a protected
call to `vim.keymap.set` for each item in the table, and display the
error before attempting to set the next keymap defined in the table.

This way, if we had a broken keymap at the top of the table, we can
still attempt to set the rest of the keymaps.