Define augroups and `User` autocmds
The plan as of now is to define the events and patterns for autocmds
created elsewhere in the configuration by creating an autocmd that
triggers a user-defined event.

This way, when we want to create an autocmd to trigger a callback, we
only need to supply a user event name along with the callback.

Pretty much we don't have to clutter up our config with pattern-matching
logic everywhere we want to create an autocmd. Instead, all that stuff
is centralized in one place so our config reads and looks a little
nicer.