WIP: utility module for `autocmds` and `augroups`
A wrapper around the API functions to manage autocommands and
autogroups. Still working on how it should expose functionality and if
it feels good to use or not.

TODO: create tests for the utility modules...

PS: I know, I know, write tests before the implementation, but I don't
know if I'll like the implementation until after I start using it. Ya
get me?