Fix some options
String-style options like "shortmess" and "formatoptions" cannot be set
with a table.

I think this is a bug, but "mousescroll" appears to be a normal map of
values but it cannot be set like it suggests in the `:h vim.opt`
documentation. At the moment, it has to be set like an array of values.