Write test for `map.iter`
The body of the test starts with declaring a local var "keymap_args" and
assigning it to an empty table. Then, we use a generic for-loop to call
`map.iter` and append "keymap_args" with a table containing the values
returned from each iteration.

After the loop, iterate through "keymap_args" and apply the validation
used in `vim.keymap.set` to each item in the list. Assert that each call
to `vim.validate` has no errors.

The local var "test_map_defs" that contains a table of mapping
definitions to be used as the argument for the "set", "iter", and
"validate" functions of the module. Its shape should be representative
of the kinds of arguments you can use when calling those functions.