Change var name from "def" to "info"
In the API documentation for the function `nvim_get_keymap`, the
description is "Gets a list of global (non-buffer-local) mapping
*definitions*" given a mode short-name. It describes the return value as
an "Array of maparg()-like dictionaries describing mappings."

In the documentation for Vim's builtin function `maparg()` describes it
as returning "a dict containing all the *information* of the mapping"
when given the `lhs`, mode, `abbrv` is false, and `dict` is true.

Regarding the semantics of "mapping definition" and "mapping
information", we will say the latter refers to the refers to the rhs and
any map arguments for a corresponding lhs in some mode. We will say the
former refers to all the details of a mapping: mode, lhs, rhs, and any
map arguments. In other words, the details required to create a mapping.