Change the shape of the "info" table
Kinda a major change...

Instead of having a list for each mapping where the first, second, and
third item is the description, rhs, and opts table respectively, have
the first item be the rhs and the description and opts table be separate
fields.

Need to change the various tables of mapping definitions in this spec
file and the `keymaps.lua` file to reflect the new required fields.