Uhhh...
Refactor to make the mapping description optional...

This kinda breaks most of the tests and I'm questioning if we really
need to have a separate validation function and other parts of the
design...

I think we should just simplify the module and remove some of the
functionality; right now I can see that it's too opinionated. I think
I'll try to leave warning messages when a mapping doesn't have a
description instead of saying they're invalid.