Implement test for calling `map.modes` in a loop
Implement a test for another iterator that only returns the modes
given either a list of modes, maparg()-like dictionaries, or a table
of mapping definitions. Access this iterator by calling `map.modes` as a
function in a generic for-loop.

Just trying this out, might just use `map.iter` to get the modes and
discard the rest of the results.