feat(nvim): utility module that wraps packer
The `util.packer` module wraps the `packer` module with functions that
configure it, bootstraps it if its not installed, etc.

The biggest addition is adding a custom key handler for every plugin
spec's "opt" key which, when true, searches for any file's in the
plugin's `lua` directory, and adds them to the "module_pattern" key. It
makes it so that any module that I require will automatically add its
respective plugin to the runtimepath.

I also have my plugin specs as submodules under "plugins". When
the parent module gets required, it calls an iterator which scans
through its enclosing directory for and looks for other Lua modules and
yields loader functions created using `loadfile`. I learned that
`require` actually uses `loadfile` under the hood, and this way we can
skip through all that path searching and setting `package.loaded` which
should make it faster than traditionally loading the submodules.

...but folke's lazy.nvim does the same thing but better so I'll just use
that instead lol