feat(nvim): `util.map` accepts trees
It works!

I like having keymaps structured as trees like which-key, but I want to
be able to do that throughout my config or in situations where I won't
have my plugins installed.

I'm thinking of deferring setting keymaps like which-key, use
`vim.keymap` to set all of them, and register any which-key specific
labels if it's installed.