feat(nvim)!: refactor keymaps to use `util.map`
Some of the callbacks for the mappings will trigger lazy to
automatically load a plugin. This is because they call `require` on a
Lua module that is provided by a plugin that lazy manages. Lazy tracks
how plugins are loaded, but in this case it cannot tell you that it was
loaded specifically because of a keymap. If we want lazy to keep track
of keymaps that load a plugin, we need to explicitly set a "keys"
property for each spec.

Different plugins have different ways of setting keymaps, and I'm hoping
that this `util.map` module will help make it more consistent.