refactor!: `util.map`
Rename `util.map.register()` to `util.map.set()` which has a slightly
different signature, logic, etc, but now returns a value which is
intended to used for a plugin specs "keys" value.

Remove the autocmd that loads all the mappings on the queue; I'm
thinking we can do that from within the "config" function that gets
executed when `which-key` gets loaded.