fix: telescope spec keys
I reread the lazy docs and it says that if you do specify a list of
LazyKeys and do not provide the rhs, the real mapping must be created in
the config function. If you don't, the mapping only gets triggered once.

I made another function in the `util.map` module called "lazykeys" which
takes the same tree of mappings as an input but returns a list of
mappings in the shape that lazy expects for a plugin spec.

I also refactored the other functions to not return anything.

That module needs to be refactored...